feature: Add support for unique WM Ids and desktop files.
Fixes #7 and #8.

I don't see much downside to this, but note it depends on the version of
Qutebrowser which I think is not yet released.

As before, there is no "delete profile" feature, and adding .desktop
files does provide one thing to manually clean up.

This work depends on #9, so it appears in this PR as well.