quit throwing away STDERR-- it provides valuable debugging information
Tonight I thought `qutebrowser-profile` was failing. Nothing happened when I ran it. 

Then I made this change locally so that STDERR was printed and `qutebrowser` shared a crucial detail: The browser window happened an opened an existing instance (on a desktop that was hidden).

Such diagnostic messages can be a big time saver so let's not throw them away.